refactor: make fetching functions importable
